# DF/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/comboboxtable/schema": {
      "get": {
        "description": "Retrive the datatype of the comboboxtable table",
        "operationId": "comboboxtableController_findSchema",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/comboboxtableEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "schema validation",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/comboboxtable/get": {
      "get": {
        "description": "Filter all the records from the comboboxtable table",
        "operationId": "comboboxtableController_findAllmethod",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of records to fetch",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/comboboxtableEntity"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Filter the records",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/comboboxtable/{my_id}": {
      "get": {
        "description": "Read only one records from the comboboxtable table",
        "operationId": "comboboxtableController_findOne",
        "parameters": [
          {
            "name": "my_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/comboboxtableEntity"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Fetch the only one record",
        "tags": [
          "ERD API"
        ]
      },
      "patch": {
        "description": "Update the record for the comboboxtable table",
        "operationId": "comboboxtableController_update",
        "parameters": [
          {
            "name": "xCdcaRole",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "xCdcaUsername",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "xCdcaRemarks",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "xCdcaApprovalStatus",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "my_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatecomboboxtableDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/comboboxtableEntity"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update the record",
        "tags": [
          "ERD API"
        ]
      },
      "delete": {
        "description": "Delete the record for the comboboxtable table",
        "operationId": "comboboxtableController_remove",
        "parameters": [
          {
            "name": "xCdcaRole",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "xCdcaUsername",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "xCdcaRemarks",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "xCdcaApprovalStatus",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "my_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/comboboxtableEntity"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete the record",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/comboboxtable": {
      "get": {
        "description": "Read all the records from the comboboxtable table",
        "operationId": "comboboxtableController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/comboboxtableEntity"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Read all the records",
        "tags": [
          "ERD API"
        ]
      },
      "post": {
        "description": "Create the record for the comboboxtable table",
        "operationId": "comboboxtableController_create",
        "parameters": [
          {
            "name": "xCdcaRole",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "xCdcaUsername",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "xCdcaRemarks",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "xCdcaApprovalStatus",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "xCdcaApprovalID",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatecomboboxtableDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/comboboxtableEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create the record",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/comboboxtable/find/first": {
      "get": {
        "description": "Read first record from the comboboxtable table",
        "operationId": "comboboxtableController_findFirst",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/comboboxtableEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Fetch the first record",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/comboboxtable/find/last": {
      "get": {
        "description": "Read last record from the comboboxtable table",
        "operationId": "comboboxtableController_findLast",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/comboboxtableEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Fetch the last record",
        "tags": [
          "ERD API"
        ]
      }
    }
  },
  "info": {
    "title": "Nest API",
    "description": "The Nestjs API documentation",
    "version": "0.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://tgadev2.toruslowcode.com/ct001/tgw/tgw004/v1/api",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "comboboxtableEntity": {
        "type": "object",
        "properties": {
          "my_id": {
            "type": "number",
            "example": "number"
          },
          "country": {
            "type": "string",
            "example": "string"
          },
          "country_code": {
            "type": "string",
            "example": "string"
          },
          "state": {
            "type": "string",
            "example": "string"
          },
          "state_code": {
            "type": "string",
            "example": "string"
          },
          "city": {
            "type": "string",
            "example": "string"
          },
          "otherdetails": {
            "type": "object",
            "example": "any"
          },
          "trs_created_date": {
            "format": "date-time",
            "type": "string",
            "example": "datetime"
          },
          "trs_created_by": {
            "type": "string",
            "example": "string"
          },
          "trs_modified_date": {
            "format": "date-time",
            "type": "string",
            "example": "datetime"
          },
          "trs_modified_by": {
            "type": "string",
            "example": "string"
          },
          "trs_process_id": {
            "type": "string",
            "example": "string"
          },
          "trs_access_profile": {
            "type": "string",
            "example": "string"
          },
          "trs_org_grp_code": {
            "type": "string",
            "example": "string"
          },
          "trs_org_code": {
            "type": "string",
            "example": "string"
          },
          "trs_role_grp_code": {
            "type": "string",
            "example": "string"
          },
          "trs_role_code": {
            "type": "string",
            "example": "string"
          },
          "trs_ps_grp_code": {
            "type": "string",
            "example": "string"
          },
          "trs_ps_code": {
            "type": "string",
            "example": "string"
          },
          "trs_sub_org_grp_code": {
            "type": "string",
            "example": "string"
          },
          "trs_sub_org_code": {
            "type": "string",
            "example": "string"
          },
          "trs_locked_by": {
            "type": "string",
            "example": "string"
          },
          "trs_locked_time": {
            "format": "date-time",
            "type": "string",
            "example": "datetime"
          },
          "trs_tenant_id": {
            "type": "string",
            "example": "string"
          },
          "trs_app_code": {
            "type": "string",
            "example": "string"
          },
          "trs_product_code": {
            "type": "string",
            "example": "string"
          },
          "trs_event_process_status": {
            "type": "string",
            "example": "string"
          },
          "trs_event_status": {
            "type": "string",
            "example": "string"
          },
          "trs_token_id": {
            "type": "string",
            "example": "string"
          },
          "trs_version": {
            "type": "string",
            "example": "string"
          }
        },
        "required": [
          "my_id"
        ]
      },
      "CreatecomboboxtableDto": {
        "type": "object",
        "properties": {
          "my_id": {
            "type": "integer",
            "format": "int32"
          },
          "country": {
            "type": "string"
          },
          "country_code": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "state_code": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "otherdetails": {
            "type": "object"
          },
          "trs_created_date": {
            "type": "string",
            "format": "date-time"
          },
          "trs_created_by": {
            "type": "string"
          },
          "trs_modified_date": {
            "type": "string",
            "format": "date-time"
          },
          "trs_modified_by": {
            "type": "string"
          },
          "trs_process_id": {
            "type": "string"
          },
          "trs_access_profile": {
            "type": "string"
          },
          "trs_org_grp_code": {
            "type": "string"
          },
          "trs_org_code": {
            "type": "string"
          },
          "trs_role_grp_code": {
            "type": "string"
          },
          "trs_role_code": {
            "type": "string"
          },
          "trs_ps_grp_code": {
            "type": "string"
          },
          "trs_ps_code": {
            "type": "string"
          },
          "trs_sub_org_grp_code": {
            "type": "string"
          },
          "trs_sub_org_code": {
            "type": "string"
          },
          "trs_locked_by": {
            "type": "string"
          },
          "trs_locked_time": {
            "type": "string",
            "format": "date-time"
          },
          "trs_tenant_id": {
            "type": "string"
          },
          "trs_app_code": {
            "type": "string"
          },
          "trs_product_code": {
            "type": "string"
          },
          "trs_event_process_status": {
            "type": "string"
          },
          "trs_event_status": {
            "type": "string"
          },
          "trs_token_id": {
            "type": "string"
          },
          "trs_version": {
            "type": "string"
          }
        },
        "required": [
          "my_id"
        ]
      },
      "UpdatecomboboxtableDto": {
        "type": "object",
        "properties": {
          "my_id": {
            "type": "integer",
            "format": "int32"
          },
          "country": {
            "type": "string"
          },
          "country_code": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "state_code": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "otherdetails": {
            "type": "object"
          },
          "trs_created_date": {
            "type": "string",
            "format": "date-time"
          },
          "trs_created_by": {
            "type": "string"
          },
          "trs_modified_date": {
            "type": "string",
            "format": "date-time"
          },
          "trs_modified_by": {
            "type": "string"
          },
          "trs_process_id": {
            "type": "string"
          },
          "trs_access_profile": {
            "type": "string"
          },
          "trs_org_grp_code": {
            "type": "string"
          },
          "trs_org_code": {
            "type": "string"
          },
          "trs_role_grp_code": {
            "type": "string"
          },
          "trs_role_code": {
            "type": "string"
          },
          "trs_ps_grp_code": {
            "type": "string"
          },
          "trs_ps_code": {
            "type": "string"
          },
          "trs_sub_org_grp_code": {
            "type": "string"
          },
          "trs_sub_org_code": {
            "type": "string"
          },
          "trs_locked_by": {
            "type": "string"
          },
          "trs_locked_time": {
            "type": "string",
            "format": "date-time"
          },
          "trs_tenant_id": {
            "type": "string"
          },
          "trs_app_code": {
            "type": "string"
          },
          "trs_product_code": {
            "type": "string"
          },
          "trs_event_process_status": {
            "type": "string"
          },
          "trs_event_status": {
            "type": "string"
          },
          "trs_token_id": {
            "type": "string"
          },
          "trs_version": {
            "type": "string"
          }
        }
      }
    }
  }
}